+++ rSYBL-control-service-pom/rSYBL-data-processing-unit/src/main/java/at/ac/tuwien/dsg/rSybl/dataProcessingUnit/monitoringPlugins/interfaces/MonitoringInterface.java
@@ -25,29 +25,29 @@ import at.ac.tuwien.dsg.csdg.elasticityInformation.ElasticityRequirement;
 
 public interface MonitoringInterface {
     
-	  public float getCpuUsage(Node node);
-	    public float getMemoryAvailable(Node node);
-	    public float getMemorySize(Node node);
-	    public float getMemoryUsage(Node node);   
-	    public float getDiskSize(Node node);
-	    public float getDiskAvailable(Node node);
-	    public float getDiskUsage(Node node);
-	    public float getCPUSpeed(Node node);
-	    public float getPkts(Node node);
-	    public float getPktsIn(Node node);
-	    public float getPktsOut(Node node);
-	    public float getReadLatency(Node node);
-	    public float getWriteLatency(Node node);
-	    public float getReadCount(Node node);
-	    public float getCostPerHour(Node node);
-	    public float getWriteCount(Node node);
-	    public float getTotalCostSoFar(Node node);
+	  public Double getCpuUsage(Node node);
+	    public Double getMemoryAvailable(Node node);
+	    public Double getMemorySize(Node node);
+	    public Double getMemoryUsage(Node node);   
+	    public Double getDiskSize(Node node);
+	    public Double getDiskAvailable(Node node);
+	    public Double getDiskUsage(Node node);
+	    public Double getCPUSpeed(Node node);
+	    public Double getPkts(Node node);
+	    public Double getPktsIn(Node node);
+	    public Double getPktsOut(Node node);
+	    public Double getReadLatency(Node node);
+	    public Double getWriteLatency(Node node);
+	    public Double getReadCount(Node node);
+	    public Double getCostPerHour(Node node);
+	    public Double getWriteCount(Node node);
+	    public Double getTotalCostSoFar(Node node);
 	    public List<String> getAvailableMetrics();
 	    public void submitServiceConfiguration(Node node);
 	    public void submitElasticityRequirements(ArrayList<ElasticityRequirement> description);
 	    public void notifyControlActionStarted(String actionName, Node node);
 		public void notifyControlActionEnded(String actionName, Node node);
-	    public float getMetricValue(String metricName, Node node);
-	    public float getNumberInstances(Node node);
+	    public Double getMetricValue(String metricName, Node node);
+	    public Double getNumberInstances(Node node);
 	    public void refreshServiceStructure(Node node);
 }
